+++ membership.php
@@ -30,6 +30,7 @@ $_TABLES['membership_trans']    = $prefix . 'trans';
 $_TABLES['membership_newusers'] = $prefix . 'newusers';
 $_TABLES['membership_log']      = $prefix . 'log';
 $_TABLES['membership_positions'] = $prefix . 'positions';
+$_TABLES['membership_posgroups'] = $prefix . 'pos_groups';
 $_TABLES['membership_users']    = $prefix . 'users';
 
 /**