+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Release TBD
 - Change Paypal plugin to Shop
 - Add a message to plan lists indicating the latest plan if expired.
 - Positions can be optionally shown on profile lists.
+- Separate position group table for more flexibility in position lists.
 
 ## 0.2.0
 - Implement PHP class autoloader